Restructure to make a github package
