+++ app/controllers/Code.scala
@@ -2,8 +2,8 @@ package controllers
 
 import javax.inject.Inject
 import play.api.mvc._
-import services.ContributorsSummariser
 import play.api.libs.concurrent.Execution.Implicits._
+import services.github.ContributorsSummariser
 
 class Code @Inject() (contributorsSummariser: ContributorsSummariser) extends Controller {
